(refactor) move spawnBot to utils folder
