+++ server/utils/spawnBot.js
